Add user related details in Cloud admin
This commit includes one addition in cloud/admin.py file and a small tweak in cloud/models.py. Now, cloud admin can get more details related to the user, like whether the user is active or a superuser. The 'user' field in the model has been updated to use a related name, 'files', for backward relations.