Changed the lookap field in authentication/views to make it easier to work with the API
Since the status and pk data of the user is not passed to the frontend for security reasons, it was not possible to request data from the client to check the admin status during authentication.

When changing a field, username is passed from the client instead of PK, which allows to conveniently send requests from the client without having to add additional checks to views.