Restrict user permissions and adjust user serialization
In the user management system, users are now prevented from deleting themselves while superusers are allowed. There is an added condition in the 'perform_destroy' method in authentication views to implement this change. Also added a new 'perform_update' method. In serializers, 'is_staff' replaced 'is_superuser' and changes in access settings are made during 'PUT' or 'PATCH' requests. This improved both security and functionality of user management and authentication.