Changed 'public_url' attribute in File model to UUID type
The 'public_url' variable in 'File' model was adjusted to utilise UUID data type instead of URL data. This change was propagated throughout the application. This allows the public access and file system to be handled in a safer and more unique manner, ensuring every file has a unique identifier when public access is ranted.