Add download date to File model upon download
The download date is now recorded in the File model whenever a file is downloaded. The Django utility now() was imported, and the File object's 'date_downloaded' attribute is set to the current time whenever the file is returned in a response. This change allows tracking of each file's download activity.