Fix superuser check in update method
The update method in the authentication views has been adjusted. The condition to check if a user is a superuser now evaluates the instance in the serializer rather than the request user. An else clause was also added to execute the regular update when the user is not a superuser.