Refine permission checks and query handling in cloud views
The code changes in `cloud/views.py` are related to setting permissions to allow administrator access to users' cloud storage and are part of the work to add the required functionality for the administrative panel.

Now the administrator can view, delete and edit user files from the administrative panel. At the same time, the administrator cannot add new files to other users' storages.