Update "date_downloaded" field in cloud models and migrations
The "date_downloaded" field in the cloud models is updated to accept null and blank entries. This change signifies that the file download time is not immediately set upon file creation, and it is only updated when the file is actually downloaded. A corresponding migration file is added to implement this change in the database schema.