Add pytest-cov package and coverage to project
Integrated pytest-cov for coverage report generation and updated python packages in the pyproject.toml and poetry.lock files. Coverage package is now included in the dependencies. Additionally, created .coveragerc file to specify omit patterns for coverage and updated Django settings to set the default format for test requests to json.