Update Django and nginx configurations
Modified url patterns in Django and adjusted upstream configurations in nginx. Changes include routing admin urls through the api path and updating the name of the upstream server in nginx. These adjustments were made for the integration of nextjs into the docker-compose setup and to ensure correct management of static and media files in Django.