Refactor Docker setup and add GitHub Actions workflow
Removed the original docker-compose file, added a server version of it with latest images, and created a new GitHub Actions workflow for Docker image building and pushing. Also, an inspection profile for commit messages was added to the vcs.xml file. The new docker-compose.server.yml refers to the latest versions of Docker images instead of building them, and the GitHub Actions workflow allows for automatic building and pushing of Docker images on changes to the deploy branch.